+++ resources/views/Consignee/Contracts/index.blade.php
@@ -42,7 +42,7 @@
                         </div>
                         <div class="card-body">
                             <h3>CONTRACT REQUEST PENDING</h3>
-                            <div style="font-size: 18px;" class="mt-4 badge badge-warning text-black disabled">
+                            <div style="font-size: 18px;" class="mt-4 badge badge-warning text-black">
                                 Waiting for Response . . .
                             </div>
 
@@ -91,7 +91,7 @@
                         </div>
                         <div class="card-body">
                             <h3>CONTRACT CHANGE REQUEST PENDING</h3>
-                            <div style="font-size: 18px;" class="mt-4 badge badge-warning text-black disabled">
+                            <div style="font-size: 18px;" class="mt-4 badge badge-warning text-black">
                                 Waiting for Response . . .
                             </div>
 
@@ -114,7 +114,7 @@
                         </div>
                         <div class="card-body">
                             <h3>WAITING FOR FINALIZATION OF CONTRACT</h3>
-                            <div style="font-size: 18px;" class="mt-4 badge badge-warning text-black disabled">
+                            <div style="font-size: 18px;" class="mt-4 badge badge-warning text-black">
                                 Waiting for Response . . .
                             </div>
                             
@@ -126,7 +126,8 @@
                 </div>
             </div>
         @elseif(($contractList[0]->enumStatus) == 'Active')
-            <div class="container">
+            <!-- Since Hindi Mo I Bibind sa button yung id para makuha yung contract ilagay natin siya sa || data-id="" || -->
+            <div class="container" id="activeContract" data-id="{{$contract[0]->intContractListID}}">
                 <div class="row">
                     <div class="col">
                         <div class="card">
@@ -134,15 +135,19 @@
                                 <h4 class="text-center text-white">{{$contract[0]->strContractListTitle}}</h4>
                             </div>
                             <div class="card-body">
-                                <div class="mt-2">
-                                    <h4 class="mt-2 text-primary">
-                                        Contract Expiry : 
-                                        <small class="ml-2 text-black">{{$contract[0]->datContractExpire}}</small>
-                                        <span class="float-right">
-                                            <span class="">Status : </span>
-                                            <button type="button" tab-index="-1" class="text-white btn btn-success btn-sm" style="font-size: 12px; border-radius: 3px; font-weight:bold; pointer-events: none;" aria-disabled="false">ACTIVE</button>
-                                        </span> 
-                                    </h4>
+                                <h4 class="text-primary">
+                                    Contract Expiry : 
+                                    <small class="ml-2 text-black">{{$contract[0]->datContractExpire}}</small>
+                                    <span class="float-right">
+                                        <span class="text-black">Status : </span>
+                                        <button type="button" tab-index="-1" class="text-white btn btn-success btn-sm" style="font-size: 12px; border-radius: 3px; font-weight:bold; pointer-events: none;" aria-disabled="true">ACTIVE</button>
+                                    </span> 
+                                </h4>
+                                <hr>
+                                <div class="row">
+                                    <div class="col-12">
+                                        <p class="text-black text-center" id="contractDetails"></p>
+                                    </div>
                                 </div>
                                 <hr>
                                 <div class="row">
@@ -151,25 +156,25 @@
                                             <li class="list-inline-item text-primary">
                                                 <p class="font-weight-bold">Standard Rate : </p></li>
                                             <li class="list-inline-item">
-                                                <p class="text-black" id="standardRate"> &#8369;</p></li>
+                                                <p class="text-black" id="standardRate"></p></li>
                                         </ul>
                                         <ul class="list-inline">
                                             <li class="list-inline-item text-primary">
                                                 <p class="font-weight-bold">Tugboat Delay Fee : </p></li>
                                             <li class="list-inline-item">
-                                                <p class="text-black" id="tugboatDelayFee"> &#8369;</p></li>
+                                                <p class="text-black" id="tugboatDelayFee"></p></li>
                                         </ul>
                                         <ul class="list-inline">
                                             <li class="list-inline-item text-primary">
                                                 <p class="font-weight-bold">Violation Fee : </p></li>
                                             <li class="list-inline-item"> 
-                                                <p class="text-black" id="violationFee"> &#8369;</p></li>
+                                                <p class="text-black" id="violationFee"></p></li>
                                         </ul>
                                         <ul class="list-inline">
                                             <li class="list-inline-item text-primary">
                                                 <p class="font-weight-bold">Consignee Late Fee : </p></li>
                                             <li class="list-inline-item">
-                                            <p class="text-black" id="consigneeLateFee"> &#8369;</p></li>
+                                            <p class="text-black" id="consigneeLateFee"></p></li>
                                         </ul>
                                     </div>
                                     <div class="col-6">
@@ -177,19 +182,19 @@
                                             <li class="list-inline-item text-primary">
                                                 <p class="font-weight-bold">Minimum Damage Fee : </p></li>
                                             <li class="list-inline-item">
-                                            <p class="text-black" id="minDamageFee"> &#8369;</p></li>
+                                            <p class="text-black" id="minDamageFee"></p></li>
                                         </ul>
                                         <ul class="list-inline">
                                             <li class="list-inline-item text-primary">
                                                 <p class="font-weight-bold">Maximum Damage Fee : </p></li>
                                             <li class="list-inline-item">
-                                                <p class="text-black" id="maxDamageFee"> &#8369;</p></li>
+                                                <p class="text-black" id="maxDamageFee"></p></li>
                                         </ul>
                                         <ul class="list-inline">
                                             <li class="list-inline-item text-primary">
                                                 <p class="font-weight-bold">Discount : </p></li>
                                             <li class="list-inline-item">
-                                                <p class="text-black" id="discount"> &#37;</p></li>
+                                                <p class="text-black" id="discount"></p></li>
                                         </ul>
                                     </div>
                                 </div>