+++ resources/views/ConsigneeAccounts/info.blade.php
@@ -1,15 +1,18 @@
 <!-- More Info Modal -->
-<div class="modal animated bounceIn fast" id="moreInfoModal" tabindex="-1" role="dialog" aria-labelledby="moreInfoModalLabel" aria-hidden="true">
-    <div class="modal-dialog modal-dialog-centered modal-lg" role="document">
+<div class="modal animated bounceIn fast" id="accountInfoModal" tabindex="-1" role="dialog" aria-labelledby="moreInfoModalLabel" aria-hidden="true">
+    <div class="modal-dialog modal-dialog modal-lg" role="document">
         <div class="modal-content">
             <div class="modal-header">
-                <div class="modal-title" id="moreInfoModalLabel"><small>Job Order # 17</small>
-                    <h4>Consignee Name</h4></div>
+                <div class="modal-title" id="moreInfoModalLabel">
+                    <small class="mb-2">Consignee Name</small>
+                    <h4 id="consigneeName" class="mt-4"></h4>
+                </div>
                 <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                     <span aria-hidden="true">&times;</span>
                 </button>
             </div>
-            <div class="modal-body">
+
+            <div class="modal-body" id="consigneeDetails">
                 <div class="row mt-2">
                     <div class="col-6">
                         <ul class="list-inline">
@@ -46,7 +49,7 @@
                         </ul>
                     </div>
                 </div>
-                <div class="row">
+                {{-- <div class="row">
                     <div class="col-12 mt-2 text-center">
                         <div class="text-primary mb-2">
                             <h4>Extra Details</h4></div>
@@ -65,11 +68,10 @@
                             </div>
                         </div>
                     </div>
-                </div>
+                </div> --}}
             </div>
             <div class="modal-footer show-on-created">
-                <button type="button" class="btn btn-secondary waves-effect" data-dismiss="modal">Close</button>
-                <button type="button" class="btn btn-primary waves-effect btnEditJO" data-toggle="modal" data-target="#JOeditModal">Edit</button>
+                <button type="button" class="btn btn-secondary waves-effect closeModalButton" data-dismiss="modal">Close</button>
             </div>
         </div>
     </div>