+++ resources/views/Joborder/index.blade.php
@@ -16,8 +16,12 @@
             </small>
             </div>
         </h1>
-        @include('Joborder.list')
-        {{-- @include('Joborder.create') --}}
+        @if(Auth::user()->enumUserType == 'Admin')
+            @include('Joborder.list')
+        @elseif(Auth::user()->enumUserType == 'Affiliates')
+            @include('Joborder.forwardlist')
+        @endif
+            {{-- @include('Joborder.create') --}}
     </section>
 @endsection